Fixes sixteenth day part two
Problem: the way we defined the within is correct, but the input data is not.
As you can see, on input data of 16th. We have ticket number bigger than 973, i.e., the greatest number of the rules. However, we can find on the input data 978, 982, and more.
Therefore, I had to change the is within method. Then everything just went as it flowed